Move existing test helpers into a new test_helpers.go file within PKI
